+++ prg/__init__.py
@@ -9,10 +9,10 @@ Example
 -------
 >>> import prg
 >>> prg.__version__
-'0.3.4'
+'0.3.5'
 """
 
-__version__ = "0.3.4"
+__version__ = "0.3.5"
 
 __all__ = [
     "__version__",